refactor: extract button
