+++ lib/widgets/sync/matrix/incoming_verification_modal.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/sync/matrix/matrix_service.dart';
+import 'package:lotti/widgets/buttons/rounded_filled_button.dart';
 import 'package:lotti/widgets/sync/matrix/verification_emojis_row.dart';
-import 'package:lotti/widgets/sync/matrix/verification_modal.dart';
 import 'package:matrix/encryption.dart';
 
 enum IncomingVerificationStep {
